backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
email-validator
aiohttp
cloudinary
requests
authlib